Generify round button and minor changes
