+++ RunningOrder/FocusableNSTextField.swift
@@ -67,3 +67,9 @@ struct FocusableNSTextFieldRepresentable: NSViewRepresentable {
         }
     }
 }
+
+struct FocusableNSTextField_Previews: PreviewProvider {
+    static var previews: some View {
+        /*@START_MENU_TOKEN@*/Text("Hello, World!")/*@END_MENU_TOKEN@*/
+    }
+}
